[REFACTOR] split GenerateCommand to GenerateEventCommand/GenerateEventMapperCommand/GenerateProjectionModelCommand.
